projects/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API"
authors = ["jacksn"]

[tool.poetry.dependencies]
python = "3.10.3"
common = {path = "../common", develop = true }
bcrypt = "3.2.0"
starlette_csrf = "1.4.3"
requests_oauthlib = "1.3.1"
httpx-oauth = "0.9.0"
fastapi = { extras = ["all"], version = "0.81.0"}
uvicorn = { extras = ["standard"], version = "0.17.6"}
fastapi-users = { extras = ["sqlalchemy"], version = "10.1.5"}

[tool.poetry.dev-dependencies]
pytest = "7.1.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"